Fix deleting post's image
-Add support to check if the image exists before trying to delete it.

+++ app/Http/Controllers/PostController.php
@@ -87,7 +87,9 @@ class PostController extends Controller
         $post = Post::findOrFail($id);
         $this->authorize('delete', $post);
         $post->delete();
-        Storage::disk('public')->delete($post->image);
+        if ($post->image != null) {
+            Storage::disk('public')->delete($post->image);
+        }
         return redirect()->route('posts.index')->with('success', 'Post deleted successfully.');
     }
